+++ src/test/java/org/scion/jpan/PackageVisibilityHelper.java
@@ -30,6 +30,7 @@ import org.scion.jpan.internal.util.IPHelper;
 import org.scion.jpan.selectors.PathSelector;
 import org.scion.jpan.selectors.PathSelectorFactory;
 import org.scion.jpan.selectors.PathSelectorFixed;
+import org.scion.jpan.selectors.PathSelectorWithRefresh;
 import org.scion.jpan.testutil.ExamplePacket;
 import org.scion.jpan.testutil.MockNetwork;
 
@@ -209,12 +210,14 @@ public class PackageVisibilityHelper {
 
   /**
    * Creates a {@link SnapScionDatagramChannel} backed by the given {@link SnapTunnelSession}.
-   * Useful for unit-testing SNAP channel behaviour without a real {@link ScionService}.
+   * Useful for unit-testing SNAP channel behavior without a real {@link ScionService}.
    */
   public static ScionDatagramChannel openSnapChannel(SnapTunnelSession session) throws IOException {
+    // TODO why do whe have this here?
+    System.err.println("FIXME ScionDatagramChannel openSnapChannel -- DEFAULT service!!!");
     DatagramChannel udp = DatagramChannel.open();
-    PathSelectorFixed selector = PathSelectorFixed.create(PathPolicy.DEFAULT);
-    PathSelectorFactory factory = PathSelectorFactory.Fixed.create(PathPolicy.DEFAULT);
+    PathSelector selector = PathSelectorWithRefresh.create(Scion.defaultService(), PathPolicy.DEFAULT);
+    PathSelectorFactory factory = PathSelectorWithRefresh.Factory.create(PathPolicy.DEFAULT);
     return new SnapScionDatagramChannel(null, udp, selector, factory, session);
   }